Item Sync and POS Item Selection

+++ app/Http/Controllers/UtilityController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Item;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 
@@ -19,9 +20,33 @@ class UtilityController extends Controller
 
             \Log::debug(request()->all());
 
-            $this->getMorphMapValue(request()->get('module'))::query()->truncate();
+            $data = null;
 
-            $data = $this->getMorphMapValue(request()->get('module'))::create(request()->get('data'));
+            switch(request()->get('module')){
+                case 'user':
+                    $this->getMorphMapValue(request()->get('module'))::query()->truncate();
+
+                    $data = $this->getMorphMapValue(request()->get('module'))::create(request()->get('data'));
+                    break;
+
+                case 'store':
+                    $this->getMorphMapValue(request()->get('module'))::query()->truncate();
+
+                    Item::query()->truncate();
+
+                    $data = $this->getMorphMapValue(request()->get('module'))::create(request()->get('data'));
+                    break;
+
+                case 'items':
+                    if ((int)request()->get('page') === 1) {
+                        $this->getMorphMapValue(request()->get('module'))::query()->truncate();
+                    }
+
+                    $this->getMorphMapValue(request()->get('module'))::insert(request()->get('data'));
+
+                    $data = ['count' => $this->getMorphMapValue(request()->get('module'))::count()];
+                    break;
+            }
 
             return response()->json([
                 'data' => $data,
